fixed some problems
now...
strrem: you can remove characters from index 0.
strsplit: if you pass a string that does not contain the separator as a parameter, it saves you an array with a single element that will contain the entire string.
general: when using the library in c++ there will be no problems of multiple definitions.